Use Releases (#2)
* Use Releases

Create full releases for tagged commits, and stop building the firmware on every commit.

* Use Generated Layout Preview in README

* Better and Fixed Path Check

* Rename Workflow and Simplify Keymap Rendering Process

The renaming of the workflow from 'draw-keymaps' to 'render-keymaps' better aligns with its actual functionality and improves understandability. Furthermore, the removal of the 'is_update_needed' job simplifies the rendering process by directly using the keymap-drawer, thereby eliminating the need to check for changes before rendering. This change aims to streamline the workflow and enhance its maintainability.

* :construction_worker: Add Release Notes

The release notes now include a visual representation of the keyboard layout changes for quick reference.

Further, the layout images are included as downloadable release assets.

The Release creation process is also updated to use a modern packaging action.

* :construction_worker: Add CODEOWNERS