:keyboard: Reorg Arrows and Management Functions (#3)
* :keyboard: Reorg Arrows and Management Functions

The keymap configuration for the Corneish Zen keyboard has been revised to improve usability and workflow. The layout of the QWERTY layer has been modified to swap the GUI and CTRL keys.

The NUMBER layer's directional keys have been reorganized for more ergonomic usage.

A new MGMT layer has been introduced, dedicated to Bluetooth device management, which enhances the ease of switching between devices.

* Update Keymap SVGs

* :construction_worker:

Work-around a bug in the caksoylar/keymap-drawer/.github/workflows/draw-zmk.yml action that made it unusable in Pull Requests. Specifically, the checkout action will present the branch in a target-merged detached head state, which is incompatible with the git-auto-commit-action action that is later used.

---------

Co-authored-by: jeffwilde <jeffwilde@users.noreply.github.com>