:keyboard: More Management Features
This commit modifies the key bindings to include system reset and bootloader controls directly in the keymap.

The MGMT layer escape (back to default layer) is also added to both halves, since the right can be reset independently, the left half needs a method to get out of MGMT mode.